// OPENPUBLICLIBRARY/OPENPUBLICLIBRARY/frontend/src/pages/catalog/bookinstance/data.d.ts
export interface BookInstancesListItemDataType {
  _id: string;
  title: string;
  cover: string;
  logo: string;
  author: string;
  first_name: string;
  family_name: string;
  date_of_birth: string;
  date_of_death: string;
  book: {};
  geren: {};
  due_back: string;
  isbn: string;
  imprint: string;
  available: number;
  status: string;
  href: string;
  summary: string;
}

// bookinstance_list: Array(14)
// 0:
// book:
// author: "5feb1a5163e28321e0935920"
// genre: Array(1)
// 0: "5feb1a5163e28321e093591c"
// lastIndex: (...)
// lastItem: (...)
// length: 1
// __proto__: Array(0)
// isbn: "0008117497"
// summary: "The Foundation series is Isaac Asimov’s iconic masterpiece. Unfolding against the backdrop of a crumbling Galactic Empire, the story of Hari Seldon’s two Foundations is a lasting testament to an extraordinary imagination, one whose unprecedented scale shaped science fiction as we know it today."
// title: "Foundation"
// __v: 0
// _id: "5feb1a5163e28321e093592a"
// __proto__: Object
// due_back: "2020-12-29T12:00:17.767Z"
// imprint: "人民文学出版社 汉日对照 2017年版"
// status: "已借出"
// __v: 0
// _id: "5feb1a5163e28321e093592c"